repos/smart-contract-tx-rollup/requirements.txt

web3==5.31.1
pytest==7.3.1
python-dotenv==0.19.2
